test(tf): fold the multi-frame DOS check into test_model_dos
Move the multi-frame global-vs-atomic-sum DOS test into TestModel in
test_model_dos.py as a second method and remove the standalone
test_model_dos_multiframe.py, so both DOS model graph tests share the same
gen_data/del_data fixture and live together.
